% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_ellipse.R
\name{confidence_ellipse}
\alias{confidence_ellipse}
\title{A Function that Computes the Confidence Ellipse Coordinates for Bivariate Normal Data (with Optional Grouping)}
\usage{
confidence_ellipse(.data, x, y, .group_by = NULL, conf_level = 0.95)
}
\arguments{
\item{.data}{The data frame or tibble.}

\item{x}{The unquoted column name for the x-axis variable.}

\item{y}{The unquoted column name for the y-axis variable.}

\item{.group_by}{The unquoted column name for the grouping variable (optional). Note that this grouping variable must be a factor.}

\item{conf_level}{The confidence level for the ellipse (0.95 by default).}
}
\value{
A data frame of the coordinates points of the ellipse.
}
\description{
This function computes a confidence ellipse (assuming bivariate normality) at a specified confidence level.
}
\examples{
# Data
data("glass", package = "ConfidenceEllipse")
# Confidence ellipse
ellipse <- confidence_ellipse(.data = glass, x = SiO2, y = Na2O)
ellipse_grp <- confidence_ellipse(
.data = glass,
x = SiO2,
y = Na2O,
.group_by = glassType
)

}
\author{
Christian L. Goueguel
}
