% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceEllipse-package.R
\docType{package}
\name{ConfidenceEllipse-package}
\alias{ConfidenceEllipse}
\alias{ConfidenceEllipse-package}
\title{ConfidenceEllipse: Confidence Ellipse Coordinates for Bivariate Normal Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package provided a function to calculate confidence ellipse, based on the assumption of bivariate normality, for a given confidence level.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://christiangoueguel.github.io/ConfidenceEllipse/}
}

}
\keyword{internal}
