% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glass}
\alias{glass}
\title{Glass Vessels Data}
\format{
Data frame of 180 rows and 14 columns.
}
\source{
\doi{10.1039/C9JA00090A}
}
\usage{
glass
}
\description{
The dataset is comprised of 13 different measurements for 180
archaeological glass vessels from different groups (Janssen, K.H.A., De Raedt, I., Schalm, O., Veeckman, J.: Microchim. Acta 15 (suppl.) (1998)
253-267. Compositions of 15th - 17th century archaeological glass vessels excavated in Antwerp.).
}
\keyword{datasets}
