% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_ellipsoid.R
\name{confidence_ellipsoid}
\alias{confidence_ellipsoid}
\title{A Function that Computes the 3D Confidence Ellipsoid Coordinates for Trivariate Normal Data (with Optional Grouping)}
\usage{
confidence_ellipsoid(.data, x, y, z, .group_by = NULL, conf_level = 0.95)
}
\arguments{
\item{.data}{The data frame or tibble.}

\item{x}{The unquoted column name for the x-axis variable.}

\item{y}{The unquoted column name for the y-axis variable.}

\item{z}{The unquoted column name for the z-axis variable.}

\item{.group_by}{The unquoted column name for the grouping variable (optional). Note that this grouping variable must be a factor.}

\item{conf_level}{The confidence level for the ellipse (0.95 by default).}
}
\value{
A data frame of the coordinates points of the ellipse.
}
\description{
A Function that Computes the 3D Confidence Ellipsoid Coordinates for Trivariate Normal Data (with Optional Grouping)
}
\examples{
# Data
data("glass", package = "ConfidenceEllipse")
# Confidence ellipsoid
ellipsoid <- confidence_ellipsoid(.data = glass, x = SiO2, y = Na2O, z = Fe2O3)
ellipsoid_grp <- confidence_ellipsoid(
.data = glass,
x = SiO2,
y = Na2O,
z = Fe2O3,
.group_by = glassType
)

}
\author{
Christian L. Goueguel
}
