% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceEllipse-package.R
\docType{package}
\name{ConfidenceEllipse-package}
\alias{ConfidenceEllipse}
\alias{ConfidenceEllipse-package}
\title{ConfidenceEllipse: Computation of 2D and 3D Confidence Ellipse Coordinates}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions designed to calculate confidence ellipses and ellipsoids at a specified confidence level.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://christiangoueguel.github.io/ConfidenceEllipse/}
  \item \url{https://github.com/ChristianGoueguel/ConfidenceEllipse}
  \item Report bugs at \url{https://github.com/ChristianGoueguel/ConfidenceEllipse/issues}
}

}
\author{
\strong{Maintainer}: Christian L. Goueguel \email{christian.goueguel@gmail.com} (\href{https://orcid.org/0000-0003-0521-3446}{ORCID})

}
\keyword{internal}
