% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceEllipse-package.R
\docType{package}
\name{ConfidenceEllipse-package}
\alias{ConfidenceEllipse}
\alias{ConfidenceEllipse-package}
\title{ConfidenceEllipse: Computation of 2D and 3D Confidence Ellipse Coordinates}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package provided a function to calculate confidence ellipses and ellipsoids for a given confidence level.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://christiangoueguel.github.io/ConfidenceEllipse/}
}

}
\keyword{internal}
