% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceEllipse-package.R
\docType{package}
\name{ConfidenceEllipse-package}
\alias{ConfidenceEllipse}
\alias{ConfidenceEllipse-package}
\title{ConfidenceEllipse: Computation of 2D and 3D Elliptical Joint Confidence Regions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Computing elliptical joint confidence regions at a specified confidence level. It provides the flexibility to estimate either classical or robust confidence regions, which can be visualized in 2D or 3D plots. The classical approach assumes normality and uses the mean and covariance matrix to define the confidence regions. Alternatively, the robustified version employs estimators like minimum covariance determinant (MCD) and M-estimator, making them less sensitive to outliers and departures from normality. Furthermore, the functions allow users to group the dataset based on categorical variables and estimate separate confidence regions for each group. This capability is particularly useful for exploring potential differences or similarities across subgroups within a dataset. Varmuza and Filzmoser (2009, ISBN:978-1-4200-5947-2). Johnson and Wichern (2007, ISBN:0-13-187715-1). Raymaekers and Rousseeuw (2019) \doi{10.1080/00401706.2019.1677270}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://christiangoueguel.github.io/ConfidenceEllipse/}
  \item \url{https://github.com/ChristianGoueguel/ConfidenceEllipse}
  \item Report bugs at \url{https://github.com/ChristianGoueguel/ConfidenceEllipse/issues}
}

}
\author{
\strong{Maintainer}: Christian L. Goueguel \email{christian.goueguel@gmail.com} (\href{https://orcid.org/0000-0003-0521-3446}{ORCID})

}
\keyword{internal}
