% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceEllipse-package.R
\docType{package}
\name{ConfidenceEllipse-package}
\alias{ConfidenceEllipse}
\alias{ConfidenceEllipse-package}
\title{ConfidenceEllipse: This Package is for Adding Confidence Ellipse to a Given 2D Scatter Plot}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://christiangoueguel.github.io/ConfidenceEllipse/}
}

}
\keyword{internal}
